+++ src/egeo-demo/barrels.ts
@@ -14,6 +14,8 @@ import { StButtonDemoModule } from './st-button-demo/st-button-demo.module';
 import { StCheckboxDemoModule } from './st-checkbox-demo/st-checkbox-demo.module';
 import { StDropdownDemoModule } from './st-dropdown-demo/st-dropdown-demo.module';
 import { StDropdownMenuDemoModule } from './st-dropdown-menu-demo/st-dropdown-menu-demo.module';
+import { StFileButtonDemoModule } from './st-file-button-demo/st-file-button-demo.module';
+
 import { StFooterDemoModule } from './st-footer-demo/st-footer-demo.module';
 import { StFormDemoModule } from './st-form-demo/st-form-demo.module';
 import { StGridDemoModule } from './st-grid-demo/st-grid-demo.module';
@@ -55,6 +57,7 @@ export const EGEO_DEMO_MODULES: any[] = [
    StCheckboxDemoModule,
    StDropdownDemoModule,
    StDropdownMenuDemoModule,
+   StFileButtonDemoModule,
    StFooterDemoModule,
    StFormDemoModule,
    StHeaderDemoModule,