+++ src/lib/st-file-button/st-file-button.module.ts
@@ -1,0 +1,11 @@
+import { CommonModule }  from '@angular/common';
+import { NgModule } from '@angular/core';
+
+import { StFileButtonComponent } from './st-file-button.component';
+
+@NgModule({
+   imports: [ CommonModule ],
+   declarations: [ StFileButtonComponent ],
+   exports: [ StFileButtonComponent ]
+})
+export class StFileButtonModule {}